talking to openai
