+++ server/graph_service/routers/ai/__init__.py
@@ -1,0 +1,3 @@
+"""
+AI module for the Graphiti server.
+"""
