+++ server/graph_service/routers/ai/config.py
@@ -1,0 +1,9 @@
+"""
+Configuration for AI services used in fact extraction.
+"""
+import os
+
+# Load settings and API key
+OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
+OPENAI_MODEL = "gpt-4.1-mini"
+TEMPERATURE = 0.25
