+++ server/graph_service/routers/ai/summary.md
@@ -1,0 +1,33 @@
+# Podsumowanie zmian dla pliku factExtractor.py
+
+## Problem
+Plik `factExtractor.py` jest zbyt duży (około 400 linii) i zawiera kilka różnych odpowiedzialności, co utrudnia:
+- zrozumienie kodu
+- debugowanie
+- testowanie
+- ponowne wykorzystanie komponentów
+
+## Proponowane rozwiązanie: podział na mniejsze moduły
+
+Proponuję podzielić plik na następujące moduły:
+
+1. **config.py** - konfiguracja OpenAI API
+2. **function_specs.py** - definicje funkcji dla OpenAI
+3. **extraction.py** - logika ekstrakcji danych przez OpenAI
+4. **neo4j_operations.py** - operacje na bazie Neo4j
+5. **factExtractor.py** (główny) - koordynator używający powyższych modułów
+
+## Korzyści
+- **Modularna struktura** - każdy moduł ma pojedynczą odpowiedzialność
+- **Łatwiejsze testy** - można testować każdy moduł niezależnie
+- **Lepsza utrzymywalność** - łatwiej wprowadzać zmiany w poszczególnych częściach
+- **Przejrzystość** - łatwiej zrozumieć, co robi każda część
+
+## Szczegółowy plan refaktoryzacji
+Szczegółowy plan, wraz z propozycją zawartości poszczególnych plików, znajdziesz w pliku `restructure.md`.
+
+## Zalecany sposób wdrożenia
+Rekomendujemy stopniowe, ewolucyjne podejście do refaktoryzacji:
+1. Utworzenie nowych plików zgodnie z proponowaną strukturą
+2. Zachowanie starej implementacji dopóki nowa nie zostanie w pełni przetestowana
+3. Stopniowe przejście na nową implementację
